functions as it should right now
date navigation now works with the Nasa Api call. the call is not instant when buttons are clicked. Search button must be clicked in order to call api. this might stay as is because there will be significantly more calls if it is done automatically every time a button is pressed.